requirements.txt
scipy==1.14.0
sympy==1.13.1
numpy==2.0.1
mpmath==1.3.0
matplotlib==3.9.1
pyqtgraph==0.13.7
pyside6==6.7.2
qtawesome==1.3.1
platypus-opt==1.2.0